Separate out the sample from the e2e tests
- added simple e2e tests + snapshots
- removed the legacy and ESM samples -- the typescript one is good enough
- moved the extra utility functions into a utils.db so that the index.ts
  remains more focussed on the usage of the library
