--- packages/ilib-scriptinfo/samples/legacy/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "ilib-scriptinfo-legacy-sample",
-  "version": "0.0.0",
-  "description": "Legacy CommonJS sample app demonstrating ilib-scriptinfo usage with ES5 syntax",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "test": "node index.js Latn"
-  },
-  "dependencies": {
-    "ilib-scriptinfo": "workspace:^"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  }
-}
